# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/branches/{branchId}/inventory": {
      "get": {
        "operationId": "InventoryController_getStockByBranch",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max items per page (1..100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (id < cursor).",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Branch inventory list"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List inventory by branch",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/branches/{branchId}/inventory/adjustments": {
      "post": {
        "operationId": "InventoryController_adjust",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "Branch id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory adjusted"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adjust inventory",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/branches/{branchId}/inventory/transfers": {
      "post": {
        "operationId": "InventoryController_transfer",
        "parameters": [
          {
            "name": "branchId",
            "required": true,
            "in": "path",
            "description": "From branch id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory transferred"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transfer inventory between branches",
        "tags": [
          "Inventory"
        ]
      }
    },
    "/api/v1/products/{productId}/stock": {
      "get": {
        "operationId": "InventoryController_getStockByProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "description": "Product id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product stock result"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get stock by product (across branches)",
        "tags": [
          "Inventory"
        ]
      }
    },
...
    "/api/v1/imports/{entity}/preview": {
      "post": {
        "operationId": "ImportsExportsController_previewImport",
        "parameters": [
          {
            "name": "entity",
            "required": true,
            "in": "path",
            "description": "Entity to import. Supported: products",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "description": "Import mode. Supported: create, update, upsert",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import preview result"
          },
          "400": {
            "description": "Invalid entity/mode/file"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview import file (validates + builds preview)",
        "tags": [
          "Imports/Exports"
        ]
      }
    },
    "/api/v1/imports/{entity}/confirm": {
      "post": {
        "operationId": "ImportsExportsController_confirmImport",
        "parameters": [
          {
            "name": "entity",
            "required": true,
            "in": "path",
            "description": "Entity to import. Supported: products",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import confirm result"
          },
          "400": {
            "description": "Invalid entity/previewId"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Confirm import preview (writes to DB)",
        "tags": [
          "Imports/Exports"
        ]
      }
    },
    "/api/v1/exports/{entity}": {
      "get": {
        "operationId": "ImportsExportsController_export",
        "parameters": [
          {
            "name": "entity",
            "required": true,
            "in": "path",
            "description": "Entity to export. Supported: products",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "columns",
            "required": false,
            "in": "query",
            "description": "Optional comma-separated list of columns to include. Can be repeated.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "required": false,
            "in": "query",
            "description": "Export format. Supported: xlsx, csv (default: xlsx)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export file. Response is a binary stream with content-disposition header."
          },
          "400": {
            "description": "Invalid entity/format/columns"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export data (returns downloadable file)",
        "tags": [
          "Imports/Exports"
        ]
      }
    }
  },
  "info": {
    "title": "Stock SaaS API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Send a valid JWT access token. This API also accepts the access token via cookie (see cookie auth)."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Optional. If present, the API will read the access token from this cookie."
      }
    },
    "schemas": {
...
      "AdjustInventoryDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "Product id (26 chars).",
            "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
            "minLength": 26,
            "maxLength": 26
          },
          "quantity": {
            "type": "number",
            "description": "Quantity delta. Can be negative or positive.",
            "example": 5
          },
          "notes": {
            "type": "string",
            "description": "Optional notes.",
            "maxLength": 500
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "TransferInventoryDto": {
        "type": "object",
        "properties": {
          "toBranchId": {
            "type": "string",
            "description": "Destination branch id (26 chars).",
            "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
            "minLength": 26,
            "maxLength": 26
          },
          "productId": {
            "type": "string",
            "description": "Product id (26 chars).",
            "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
            "minLength": 26,
            "maxLength": 26
          },
          "quantity": {
            "type": "number",
            "description": "Quantity to transfer (>= 1).",
            "example": 2,
            "minimum": 1
          },
          "notes": {
            "type": "string",
            "description": "Optional notes.",
            "maxLength": 500
          }
        },
        "required": [
          "toBranchId",
          "productId",
          "quantity"
        ]
      },
...
      "ConfirmImportDto": {
        "type": "object",
        "properties": {
          "previewId": {
            "type": "string",
            "description": "Preview id returned by the preview endpoint (26 chars).",
            "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
            "minLength": 26,
            "maxLength": 26
          }
        },
        "required": [
          "previewId"
        ]
      }
    }
  }
}