# src/imports-exports/dto/confirm-import.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, Length } from 'class-validator';

export class ConfirmImportDto {
  @ApiProperty({
    description: 'Preview id returned by the preview endpoint (26 chars).',
    example: '01J1QZQ0VQ8J7TQH0YV3A1BCDE',
    minLength: 26,
    maxLength: 26,
  })
  @IsString()
  @Length(26, 26)
  previewId!: string;
}